backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-SocketIO==5.3.4
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
python-dotenv==1.0.0
eventlet==0.33.3